ffmpeg/qsv: move HEVC10 Encoder Test classes to lib
